Add program date time tag without functioning parser
NSISO8601DateFormatter, which I had been counting on for date parsing,
is not well-suited to the task. It has to be configured with the
components of the date-time it expects to encounter, which is not
feasible for HLS parsing, as only ISO/IEC 8601:2004 is specified and
time zone and fractional parts of seconds are merely suggested.