Add value equality to tags
Also adds a round-trip parse, serialize, parse test. We don't check that
the serialized playlists are equal to the original, but we do test that
both parsed lists are equal. Notably, when serializing we don't care to
preserve the order of attribute list parameters nor the presence of
default values.